refactoring for Titanium SDK 6.0.0.GA
