+++ .classpath
@@ -1,17 +1,16 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <classpath>
-	<classpathentry kind="src" path="src"/>
-	<classpathentry kind="src" path="build/.apt_generated"/>
+	<classpathentry kind="src" path="android/src"/>
+	<classpathentry kind="src" path="android/build/.apt_generated"/>
 	<classpathentry kind="con" path="org.eclipse.jdt.launching.JRE_CONTAINER"/>
-	<classpathentry kind="lib" path="/usr/local/opt/android-sdk/platforms/android-8/android.jar"/>
-	<classpathentry kind="lib" path="/usr/local/opt/android-sdk/add-ons/addon-google_apis-google_inc_-8/libs/maps.jar"/>
-	<classpathentry kind="lib" path="/Users/yellow/Library/Application Support/Titanium/mobilesdk/osx/3.0.2.GA/android/titanium.jar"/>
-	<classpathentry kind="lib" path="/Users/yellow/Library/Application Support/Titanium/mobilesdk/osx/3.0.2.GA/android/js.jar"/>
-	<classpathentry kind="lib" path="/Users/yellow/Library/Application Support/Titanium/mobilesdk/osx/3.0.2.GA/android/kroll-common.jar"/>
-	<classpathentry kind="lib" path="/Users/yellow/Library/Application Support/Titanium/mobilesdk/osx/3.0.2.GA/android/kroll-apt.jar"/>
-	<classpathentry kind="lib" path="lib/gcm.jar"/>
-	<classpathentry kind="lib" path="/Users/yellow/Library/Application Support/Titanium/mobilesdk/osx/3.0.2.GA/android/modules/titanium-android.jar"/>
-	<classpathentry kind="lib" path="/Users/yellow/Library/Application Support/Titanium/mobilesdk/osx/3.0.2.GA/android/android-support-v4.jar"/>
+	<classpathentry kind="lib" path="/Users/ingrowth/Library/android-sdk-macosx/platforms/android-23/android.jar"/>
+	<classpathentry kind="lib" path="/Users/ingrowth/Library/android-sdk-macosx/add-ons/addon-google_apis-google-23/libs/maps.jar"/>
+	<classpathentry kind="lib" path="/Users/ingrowth/Library/Application Support/Titanium/mobilesdk/osx/6.0.0.GA/android/titanium.jar"/>
+	<classpathentry kind="lib" path="/Users/ingrowth/Library/Application Support/Titanium/mobilesdk/osx/6.0.0.GA/android/kroll-common.jar"/>
+	<classpathentry kind="lib" path="/Users/ingrowth/Library/Application Support/Titanium/mobilesdk/osx/6.0.0.GA/android/kroll-apt.jar"/>
+	<classpathentry kind="lib" path="android/lib/google-play-services.jar"/>
+	<classpathentry kind="lib" path="/Users/ingrowth/Library/Application Support/Titanium/mobilesdk/osx/6.0.0.GA/android/modules/titanium-android.jar"/>
+	<classpathentry kind="lib" path="/Users/ingrowth/Library/Application Support/Titanium/mobilesdk/osx/6.0.0.GA/android/android-support-v4.jar"/>
 	<classpathentry kind="src" path=".apt_generated">
 		<attributes>
 			<attribute name="optional" value="true"/>
