+++ android/build.properties
@@ -1,0 +1,4 @@
+titanium.platform=/Users/ingrowth/Library/Application Support/Titanium/mobilesdk/osx/6.0.0.GA/android
+android.platform=/Users/ingrowth/Library/android-sdk-macosx/platforms/android-23
+google.apis=/Users/ingrowth/Library/android-sdk-macosx/add-ons/addon-google_apis-google-23
+android.ndk=/Users/ingrowth/Library/android-sdk-macosx/ndk-bundle
