+++ android/documentation/index.md
@@ -1,0 +1,13 @@
+# gcm.js
+
+**gcm.js** is an open source module for Titanium Android SDK that lets developers receive GCM push notifications in their Android apps.
+It has a very simple API -almost identical to iOS!- yet flexible and powerful, as it executes Javascript whenever a notification is received, no matter if the app is in foreground or background. More info at:
+[http://iamyellow.net/post/40100981563/gcm-appcelerator-titanium-module](http://iamyellow.net/post/40100981563/gcm-appcelerator-titanium-module)
+
+## Author
+
+jordi domenech, [iamyellow.net](http://iamyellow.net)
+
+## License
+
+Apache License, Version 2.0
