+++ android/hooks/README
@@ -1,0 +1,1 @@
+These files are not yet supported as of 1.4.0 but will be in a near future release.
