+++ android/java-sources.txt
@@ -1,0 +1,7 @@
+"/Users/ingrowth/Documents/Projects/Ti-Android-gcm.js/android/src/net/iamyellow/gcmjs/AppStateListener.java"
+"/Users/ingrowth/Documents/Projects/Ti-Android-gcm.js/android/src/net/iamyellow/gcmjs/GCMBroadcastReceiver.java"
+"/Users/ingrowth/Documents/Projects/Ti-Android-gcm.js/android/src/net/iamyellow/gcmjs/GCMIntentService.java"
+"/Users/ingrowth/Documents/Projects/Ti-Android-gcm.js/android/src/net/iamyellow/gcmjs/GcmjsActivity.java"
+"/Users/ingrowth/Documents/Projects/Ti-Android-gcm.js/android/src/net/iamyellow/gcmjs/GcmjsModule.java"
+"/Users/ingrowth/Documents/Projects/Ti-Android-gcm.js/android/src/net/iamyellow/gcmjs/GcmjsService.java"
+"/Users/ingrowth/Documents/Projects/Ti-Android-gcm.js/android/build/generated/java/net/iamyellow/gcmjs/GcmjsBootstrap.java"
