--- build.properties
@@ -1,3 +1,0 @@
-titanium.platform=/Users/tadatoshi_hanazaki/Library/Application Support/Titanium/mobilesdk/osx/5.0.2.GA/android
-android.platform=/Users/tadatoshi_hanazaki/Library/android-sdk-macosx/platforms/android-22
-google.apis=/Users/tadatoshi_hanazaki/Library/android-sdk-macosx/add-ons/addon-google_apis-google-22
