+++ cp_zip.sh
@@ -1,5 +1,3 @@
 #!/bin/bash
 
-cp dist/net.iamyellow.gcmjs-android-*.zip ./example/
-
-
+cp android/dist/net.iamyellow.gcmjs-android-*.zip ./example/
