+++ example/plugins/ti.alloy/hooks/alloy.js
@@ -5,6 +5,7 @@
  */
 
 exports.cliVersion = '>=3.X';
+exports.version = '1.0.0';
 var SILENT = true;
 
 exports.init = function (logger, config, cli, appc) {
@@ -19,9 +20,9 @@ exports.init = function (logger, config, cli, appc) {
 		spawn = require('child_process').spawn,
 		parallel = appc.async.parallel;
 
-		if(!process.env.sdk) {
-			process.env.sdk = cli.sdk.name;
-		}
+	if (!process.env.sdk) {
+		process.env.sdk = cli.sdk.name;
+	}
 
 	function run(deviceFamily, deployType, target, finished, silent) {
 		var appDir = path.join(cli.argv['project-dir'], 'app');
@@ -49,7 +50,7 @@ exports.init = function (logger, config, cli, appc) {
 				deploytype: deployType || cli.argv['deploy-type'] || 'development',
 				target: target
 			};
-		if(silent) {
+		if (silent) {
 			// turn off all logging output for code analyzer build hook
 			config.noBanner = 'true';
 			config.logLevel = '-1';
@@ -83,6 +84,7 @@ exports.init = function (logger, config, cli, appc) {
 			// we have no clue where alloy is installed, so we're going to subprocess
 			// alloy and hope it's in the system path or a well known place
 			var paths = {};
+			var locatorCmd = process.platform === 'win32' ? 'where' : 'which';
 			parallel(this, ['alloy', 'node'].map(function (bin) {
 				return function (done) {
 					var envName = 'ALLOY_' + (bin === 'node' ? 'NODE_' : '') + 'PATH';
@@ -90,11 +92,11 @@ exports.init = function (logger, config, cli, appc) {
 					paths[bin] = process.env[envName];
 					if (paths[bin]) {
 						done();
-					} else if (process.platform === 'win32') {
+					} else if (process.platform === 'win32' && bin === 'alloy') {
 						paths.alloy = 'alloy.cmd';
 						done();
 					} else {
-						exec('which ' + bin, function (err, stdout, strerr) {
+						exec(locatorCmd + ' ' + bin, function (err, stdout, strerr) {
 							if (!err) {
 								paths[bin] = stdout.trim();
 								done();
@@ -116,16 +118,15 @@ exports.init = function (logger, config, cli, appc) {
 					}
 				};
 			}), function () {
+
+				// compose alloy command execution
 				var cmd = [paths.node, paths.alloy, 'compile', appDir, '--config', config];
 				if (cli.argv['no-colors'] || cli.argv['color'] === false) { cmd.push('--no-colors'); }
-				if (process.platform === 'win32') { cmd.shift(); }
-				logger.info(__('Executing Alloy compile: %s', cmd.join(' ').cyan));
 
-				var child = (process.platform === 'win32') ? spawn(cmd.shift(), cmd, { stdio: 'inherit' }) : spawn(cmd.shift(), cmd);
-
+				// process each line of output from alloy
 				function checkLine(line) {
 					var re = new RegExp(
-						'(?:\u001b\\[\\d+m)?\\[?(' +
+						'^(?:\u001b\\[\\d+m)?\\[?(' +
 						logger.getLevels().join('|') +
 						')\\]?\s*(?:\u001b\\[\\d+m)?(.*)', 'i'
 					);
@@ -139,48 +140,55 @@ exports.init = function (logger, config, cli, appc) {
 					}
 				}
 
-				child.stdout !== null && child.stdout.on('data', function (data) {
-					data.toString().split('\n').forEach(function (line) {
-						checkLine(line);
+				// execute alloy in os-specific manner
+				var child;
+				if (process.platform === 'win32' && paths.alloy === 'alloy.cmd') {
+					cmd.shift();
+					logger.info(__('Executing Alloy compile: %s',
+						['cmd', '/s', '/c'].concat(cmd).join(' ').cyan));
+
+					// arg processing from https://github.com/MarcDiethelm/superspawn
+					child = spawn('cmd', [['/s', '/c', '"' +
+						cmd.map(function(a) {
+							if (/^[^"].* .*[^"]/.test(a)) return '"' + a + '"'; return a;
+						}).join(' ') + '"'].join(' ')], {
+							stdio: 'inherit',
+							windowsVerbatimArguments: true
+						}
+					);
+				} else {
+					logger.info(__('Executing Alloy compile: %s', cmd.join(' ').cyan));
+					child = spawn(cmd.shift(), cmd);
+					child.stdout.on('data', function (data) {
+						data.toString().split('\n').forEach(checkLine);
 					});
-				});
-				child.stderr !== null && child.stderr.on('data', function (data) {
-					data.toString().split('\n').forEach(function (line) {
-						checkLine(line);
+					child.stderr.on('data', function (data) {
+						data.toString().split('\n').forEach(checkLine);
 					});
-				});
-				child !== null && child.on('exit', function (code) {
+				}
+
+				// handle the completion of alloy, success or otherwise
+				child.on('exit', function (code) {
 					if (code) {
 						logger.error(__('Alloy compiler failed'));
 						process.exit(1);
 					} else {
 						logger.info(__('Alloy compiler completed successfully'));
+
+						afs.exists(path.join(cli.argv['project-dir'], 'build', 'i18n')) && process.argv.push('--i18n-dir', 'build');
+						afs.exists(path.join(cli.argv['project-dir'], 'build', 'platform')) && (cli.argv['platform-dir'] = 'build/platform');
 					}
 					finished();
 				});
+
 			});
 		}
 	}
 
 	cli.addHook('build.pre.compile', function (build, finished) {
-		// TODO: Remove this workaround when the CLI reports the right deploy type for android
-		var deployType = build.deployType;
-		var target = build.target;
+		var deployType = build.deployType,
+			target = build.target;
 
-		if (cli.argv.platform === 'android') {
-			switch(target) {
-				case 'dist-playstore':
-					deployType = 'production';
-					break;
-				case 'device':
-					deployType = 'test';
-					break;
-				case 'emulator':
-				default:
-					deployType = 'development';
-					break;
-			}
-		}
 		run(build.deviceFamily, deployType, target, finished);
 	});
