--- manifest
@@ -1,19 +1,0 @@
-#
-# this is your module manifest and used by Titanium
-# during compilation, packaging, distribution, etc.
-#
-version: 1.0
-apiversion: 3
-description: Simple yet powerful API for Android GCM push notifications
-author: jordi domenech
-license: Apache License, Version 2.0
-copyright: Copyright (c) 2013 by iamyellow.net
-
-
-# these should not be edited
-name: gcmjs
-moduleid: net.iamyellow.gcmjs
-guid: 
-platform: android
-minsdk: 6.0.0.GA
-architectures: armeabi-v7a x86
