--- src/net/iamyellow/gcmjs/GcmjsActivity.java
@@ -1,35 +1,0 @@
-//
-//   Copyright 2013 jordi domenech <http://iamyellow.net, jordi@iamyellow.net>
-//
-//   Licensed under the Apache License, Version 2.0 (the "License");
-//   you may not use this file except in compliance with the License.
-//   You may obtain a copy of the License at
-//
-//       http://www.apache.org/licenses/LICENSE-2.0
-//
-//   Unless required by applicable law or agreed to in writing, software
-//   distributed under the License is distributed on an "AS IS" BASIS,
-//   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-//   See the License for the specific language governing permissions and
-//   limitations under the License.
-//
-
-package net.iamyellow.gcmjs;
-
-import org.appcelerator.titanium.TiApplication;
-import ti.modules.titanium.android.TiJSActivity;
-
-public class GcmjsActivity extends TiJSActivity {
-
-	public GcmjsActivity() {
-		super("gcm_activity.js");
-
-		if (TiApplication.getAppCurrentActivity() == null) {
-			GcmjsModule.logd("Starting gcm_activity.js, app is NOT running.");
-			AppStateListener.appWasNotRunning = true;
-		} else {
-			GcmjsModule.logd("Starting gcm_activity.js, app IS running.");
-			AppStateListener.appWasNotRunning = false;
-		}
-	}
-}
