--- timodule.xml
@@ -1,39 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<ti:module xmlns:ti="http://ti.appcelerator.org" xmlns:android="http://schemas.android.com/apk/res/android">
-	<android xmlns:android="http://schemas.android.com/apk/res/android">
-		<manifest>
-			<application>
-				<!-- begin gcm.js module -->
-				<receiver android:name="net.iamyellow.gcmjs.GCMBroadcastReceiver" 
-					android:permission="com.google.android.c2dm.permission.SEND">
-					<intent-filter>
-						<action android:name="com.google.android.c2dm.intent.RECEIVE" />
-						<action android:name="com.google.android.c2dm.intent.REGISTRATION" />
-						<category android:name="${tiapp.properties['id']}" />
-					</intent-filter>
-				</receiver>
-				<service android:name="net.iamyellow.gcmjs.GCMIntentService" />
-				<service android:name="net.iamyellow.gcmjs.GcmjsService"/>
-				<activity android:name="net.iamyellow.gcmjs.GcmjsActivity" android:noHistory="true" android:theme="@style/Theme.Titanium">
-					<intent-filter>
-						<action android:name="android.intent.action.VIEW"/>
-						<category android:name="android.intent.category.DEFAULT"/>
-						<category android:name="android.intent.category.BROWSABLE"/>
-					</intent-filter>
-				</activity>
-			</application>
-
-			<!-- begin gcm.js module -->
-			<permission android:name="${tiapp.properties['id']}.permission.C2D_MESSAGE" android:protectionLevel="signature" />
-			<uses-permission android:name="${tiapp.properties['id']}.permission.C2D_MESSAGE" />
-			<uses-permission android:name="com.google.android.c2dm.permission.RECEIVE" />
-			<uses-permission android:name="android.permission.INTERNET" />
-			<uses-permission android:name="android.permission.GET_ACCOUNTS" />
-			<uses-permission android:name="android.permission.WAKE_LOCK" />
-
-			<uses-permission android:name="android.permission.GET_TASKS" />
-			<uses-permission android:name="${tiapp.properties['id']}.c2dm.permission.RECEIVE" />
-			<!-- end gcm.js module -->
-		</manifest>
-	</android>
-</ti:module>
